8.13. Оптимизируй это (часть 2): Создан обработчик handleChange отвечающий за изменение полей формы компонента ReviewForm. Обработчики submit и change для формы и её элементов перенесены в компонент ReviewForm. Сам компонент переделан во имя читабельности
